pop up bugs fixed

+++ MM-app/lib/BudgetSimulator/Widgets/overallScoreScreen.dart
@@ -28,7 +28,7 @@ class OverallScoreScreen extends StatefulWidget {
 
 class _OverallScoreScreenState extends State<OverallScoreScreen> {
   BudgetSimulatorFunctions functions = BudgetSimulatorFunctions();
-  int score = 0;
+  double score = 0;
   Color textColor = Color.fromRGBO(0, 0, 0, 0);
   String grade = '';
   String description = '';
@@ -40,7 +40,7 @@ class _OverallScoreScreenState extends State<OverallScoreScreen> {
       score += functions
           .s1P2SavingsScore(widget.widget.widget.savingsAccountBalance);
       score += functions.s1P2CreditScoreScore(widget.widget.widget.creditScore);
-      score += widget.widget.widget.onTimePaymentScore as int;
+      score += widget.widget.widget.onTimePaymentScore.toInt();
       score += functions.s1P2CCDebtScore(widget.widget.widget.creditCardDebt);
       score += functions.s1P2WellnessScore(widget.widget.widget.bodyScore,
           widget.widget.widget.mindScore, widget.widget.widget.socialScore) as int;
